Emit upstream Installation instructions (#2068)
This pull request is the scaffold implementation of generating
installation docs for bridged providers from upstream source.

In this change, a new "language" called `registry-docs` is added to the
bridge. (Note that "language" in this case is a bit of a misnomer - it
is the `tfgen` CLI's input for determining which generating tactic to
use - this is why `schema` is a language as well).

This change reads the upstream `index.md` file and emits it as a Pulumi
[package installation instruction
file](https://www.pulumi.com/docs/using-pulumi/pulumi-packages/how-to-author/#overview-installation--configuration)
to a top-level folder called `docs/` (the current registry standard). We
also emit a minimal `_index.md` file. When these files exist in a
provider, the registry will use those instead of the hard coded files
existing in the registry currently.

This pull request includes a fair bit of TODOs in the docs parser -
currently only the Hugo front matter is implemented, to show the
scaffold. Transformations for this document to be Pulumi-friendly will
follow in a separate PR.

This change does not affect any current use of the bridge - to flip it
on, we'd add another Make target to a provider's Makefile, e.g.
`./bin/pulumi-tfgen-openstack registry-docs --out docs/`. I'd like us to
consider adding this change despite the TODOs, so we can keep the
changes somewhat atomic and reduce merge conflicts.

Closes https://github.com/pulumi/pulumi-terraform-bridge/issues/2082.