Upgrade Pulumi to v3.120.0 (simply) (#2112)
Given the difficulties associated with #2109 and #2099, I thought we
should try doing the minimal work necessary to upgrade. We can then
iterate from here without falling further behind on pulumi/pulumi.

Update 1: I believe that `TestCollectionsNullEmptyRefreshClean`
continues to fail. This is unfortunate but not shocking given that it is
an integration test on refresh, and we changed the behavior of refresh
in pulumi/pulumi v3.120.0. It looks like that is the only test failing.
`make -C pf test` runs cleanly.

Update 2: The test resolves when you upgrade the local CLI for `pulumi`
to v3.120.0.