Fix bridge not running state upgrades from 0 -> 1 (#2081)
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2039


The bridge assumes that a missing `schema_version` in the meta state
means "current version". That is wrong as we only miss the
`schema_version` if it is 0 - we write all non-zero versions to the
state.

Any resources created under a v0 TF schema version will not save the the
v0 in the state. This means that if this resource then gets a TF state
upgrade, we will assume we are already on the new version and NOT run
the state upgrade as we should.

This PR changes the assumption to mean missing `schema_version` means
"v0". Note that `TestUpgradeInputsStringBasicNonZeroVersionSame` passed
before, which shows that any existing resources which had state upgrades
defined should have the correct state version in their meta property.

It also adds tests around this both for PRC and non-PRC.


The feature is flagged behind a provider flag, supplied in
`ProviderInfo` - the feature flag mechanism is adapted from the sdkv2
feature flag mechanism.