feat: optimization of nested description parsing (#2151)
This pull request adds additional docs parsing capability for type
descriptions.

Prior to this change, we match the type object key, with the the
best-attempt docs paths for nested schemas from the docs parser
Arguments, via GetNestedDescriptionsFromParsedDocs. If we do not find
any Description via that, then we attempt to obtain a Description from
the schema Description field.

With this change, when we have an empty Description, we look again at
the parsed entityDocs, and **if** the leaf of the type's docsPath is
unique within the resource, then we match that Description to the Type.

We do this matching _after_ the schema description lookup to scope this
optimization as tightly as possible - when, for instance, a field is
entirely missing from the entityDocs.Attributes, we don't want to
accidentally be matching a type description with the same docsPath leaf.
The chance of this happening is greatly reduced when checking all other,
reliable, sources first.

On AWS, this adds ~1400 descriptions to type fields:

```
git diff master --stat -- provider/cmd/pulumi-resource-aws/schema.json 
 provider/cmd/pulumi-resource-aws/schema.json | 4272 ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++---------------------------------
 1 file changed, 2809 insertions(+), 1463 deletions(-)
 ```

Fixes #1886.