Fix nested computed read (#2181)
Fixes an issue with the SDKv2 and PF input extraction. During import we
wrongly return values for properties which are fully computed which
causes invalid programs to be generated.

~The object check in `CastToTypeObject` was wrong for SDKV2 objects:
https://github.com/pulumi/pulumi-terraform-bridge/blob/48fd41bd16e2f7f933bb9390ab7619d52faabb6d/pkg/tfshim/util/types.go#L36~

The `CastToTypeObject` object check is wrong for both SDKv2 and PF. For
the SDKV2 only sets and lists can have a `Resource` `.Elem`. For PF it
list-nested blocks are represented with a `Resource` `.Elem` in the shim
layer, so the check did not catch these. PF details here:
https://github.com/pulumi/pulumi-terraform-bridge/pull/2231
This fixes the check and adds regression tests for the broken imports
which resulted from 2180.

Test coverage added in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2224 for sdkv2
and https://github.com/pulumi/pulumi-terraform-bridge/pull/2225 for pf.
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2180
depends on https://github.com/pulumi/providertest/pull/91
stacked on https://github.com/pulumi/pulumi-terraform-bridge/pull/2225