fix(pf): remove __defaults entry from inputs
PF resources should not be injecting __defaults during
import as they do not support it, but it is happening through an
unintended side-effect of code reuse from sdkv2.

I could think of a couple other ways to handle this:

1. Duplicate the code in the pf and remove the code specific to
injecting __defaults
2. Removing where we inject an empty __defaults. See #2236

Decided to go with the simplest approach of just doing some post
processing since touching the `sdkv2` code seems to be a much bigger
effort.

fixes #2189, fixes pulumi/pulumi-aws#4204