feat: validate that pf resources do not have an invalid "id" input property
This modifies the checks to account for cases where the provider has an
input property named "id" which has not been remapped to a different
name.

The linked issue mentions needing to account for sdkv2 resources, but
the existing checks skip sdkv2 resources saying that sdkv2 resources
always have an (i'm assuming valid) ID. If it is possible to have an
invalid "id" input property for sdkv2 resources then I can modify to
account for that case.

closes #2084