Fix PRC big int creation (#2286)
Big integers get truncated under PRC, even if treated as strings on the
pulumi side. This means that
https://github.com/pulumi/pulumi-terraform-bridge/pull/1896 does not
work under PRC.

The reason for that is that we do a round trip to JSON when creating the
resource where we truncate the integer. When decoding json numbers we
have to use json.Number instead of the default float in order to not
truncate 64 bit integers.

Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2284