Generate shorter names for properties
- Most of the `schemaNestedTypes` code was just moved from
  `generate_schema.go`
- The new code is the code related to `nestedTypeGraph`

I picked an arbitrary max character limit of 120, but those still look
pretty long. I checked in `pulumi-aws` and there are still a lot of
types that are up to and over 120.

I currently just pick the shorter name if the normal name would be +120,
but I feel like it would be better to introduce a new schema property to
enable this at the resource level.

TODO:
 - I need to add some more tests to ensure this is deterministic, but I
   am not yet sure it is possible to make it completely deterministic
   without tracking state (maybe by reading in the existing schema.json)

re #1118