Skip sections by header with goldmark (#2292)
This pull request introduces a Markdown transformer that can in
conjunction with a parser/renderer remove unwanted sections from a
markdown document by screening for words in the section header. This
functionality is aimed at and implemented for the
installation-configuration document, but it can be used as a
docsEditRule by any provider on any file name. It is implemented as a
docsEditRule in this pull request as well.

Note: A test was accidentally using writeFile instead of assert; this is
fixed.
Additionally, the tests verify whether transformed results are equal in
HTML, to account for parse idiosyncracies.

Resolves https://github.com/pulumi/pulumi-terraform-bridge/issues/2227.