feat: run validate during import and drop invalid inputs
> This is a reroll of #2277 which was reverted in #2299
> I've added additional handing for secret values

During an import operation property values are read from the cloud and
turned into inputs which then becomes part of the generated code. For
some properties (Computed & Optional) the cloud can return values that
fail validation. Since these are computed & optional the provider will
allow the user to not specify a value. In these cases the import would
succeed, but if the user took the generated code as is and attempted to
run pulumi up it would fail due to the validation errors.

Rather than generate invalid inputs we will instead run Validate during import
and then processes any resulting validation errors. If any properties
fail validation (and they match our other cases) then we remove those
inputs so that the generated program code will not contain the invalid
inputs

NOTE: we want to (for now at least) scope this as narrowly as possible because we do not want to
create a scenario where the resulting inputs create a different error
(e.g. removing an invalid required property). In these cases we will just allow the invalid configuration
through and the user will have to decide what to do.

fixes pulumi/pulumi-aws#4054