Make regex link matching more strict (#2344)
While rolling out https://github.com/pulumi/home/issues/3598 on Datadog,
I noticed that the default edit rules stripped the following upstream
text
>The [Datadog](https://www.datadoghq.com) provider is used to interact
with the resources supported by Datadog. The provider needs to be
configured with the proper credentials before it can be used. It
requires terraform 0.12 or later.
>Try the [hands-on
tutorial](https://developer.hashicorp.com/terraform/tutorials/use-case/datadog-provider)
on the Datadog provider on the HashiCorp Learn site.

to render

"The Datadog on the Datadog provider on the HashiCorp Learn site."

This is because the current regex matches _all_ characters before the
`terraform` link string, leading to a nonsensical replacement of the
entire match. With this change, we exclude `)` which selects for the
matching link only.

Added a test to this effect.