Implement Dynamic type support (#2362)
This PR adds support for the [Dynamic
Type](https://developer.hashicorp.com/terraform/plugin/framework/handling-data/types/dynamic#)
to Plugin Framework based bridged providers.

One use case for these types is modeling K8S manifests. In HCL this
permits users to specify manifests as verbatim trees. This PR makes it
possible in the Pulumi projection as well.

Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2127

Limitations:

- will fail if nested secrets are used inside a property of type Dynamic
- will fail if Assets, Archives, or Resource References are passed down
through a property of type Dynamic
- there is a little guess-work involved in recovering the type of the
passed value which might result in providers receiving a different
recovered type under Pulumi compared to TF; specifically maps and
objects may be confused, as can arrays and tuples; for most code this is
not expected to be a problem, but might need to be revisited if users
report issues

Dependencies: Plugin Framework updated to 1.7.0 to be able to test
dynamic types (introduced in that version).

---------

Co-authored-by: Ian Wahbe <me@iwahbe.com>