Fixup now renames "id" fields when they are user settable
`fixup` already renames fields called `"urn"`, since it is always invalid for a provider
resource to have a property called `"urn"`, since Pulumi will override it.

`fixup` now renames fields called `"id"` *only* when the `"id"` field is user
settable (Optional || Required), since Pulumi does not respect user set Inputs.

When we expect to have an ID property available (because the user or the provider will set
it), we use that as the source of our computed ID. If not, we still fallback to using
"missing ID" as the computed ID.

Fixes https://github.com/pulumi/pulumi-terraform-provider/issues/29