Render tables back to markdown
The markdown renderer we use doesn't handle tables. Instead, they were rendered as HTML,
which is fine for the registry but a poor experience for the docs embedded into our
SDKs. This commit provides a custom renderer for markdown tables, based on
"github.com/olekukonko/tablewriter" for the actual table formatting.