Custom timeouts cross test (#2391)
This adds cross tests for custom timeouts, comparing our handling of
custom timeouts with terraform's. To do that we've exposed the `plan`
error in tfcheck.

Looks like the TF rules are the following:
- If the schema specifies a timeout for the operation, then the user is
allowed to override that.
- If the schema does not specify a timeout, then the resource is said
not to support timeouts. That makes it a runtime error if the user then
tries to customise the timeout on the resource.

That makes sense conceptually, as not all resources would be implemented
in a way which supports timeouts.

Related to https://github.com/pulumi/pulumi-terraform-bridge/issues/2386