Pass `PlannedPrivate` from PlanResourceChange to ApplyResourceChange
The `PlannedPrivate` field of the PRC response includes metadata
that's required for correctly re-assembling the diff.
For example, it includes `NewExtra` which includes the original
value for attributes with a `StateFunc`.
Without `PlannedPrivate`, those attributes get saved to state
encoded twice. This can lead to mangled data in state and
perma-diffs.