Fallback to displaying the TF diff when no Pulumi diff is available
When the processes a Diff request from the Pulumi engine, it queries the Terraform
provider, receiving a `shim.DiffInstance` in response. The `shim.DiffInstance` sits in the
Terraform namespace. The bridge then tries to translate it back into the Pulumi namespace.

Most of the time, this is successful, and the user sees high quality details paths in
their diff display. In rare cases, we fail to translate the diff and the user sees that a
resource needs updating, but not which fields caused the update.

This PR adds a backstop for the above scenario:

If Pulumi is not able to specify which fields have changed but the Terraform diff does
display fields that have changed (and the resource will be updated/replaced), then the
bridge will log at `info` level the terraform diff and a suggestion to report the bug to
the Pulumi provider or the bridge.

If the new logic is activated, that means TF specified a diff with specific fields **and**
the bridge failed to translate that diff.

---

Q: Why do we need to log this info, can we not just put it in the detailed diff?

A: The detailed diff protocol only allows the provider to specify paths (in Pulumi
space). Values diffs are calculated in the engine. For TF diffs where we are unable to
translate to Pulumi paths, we can't use detailed diff.