Add SDKv2 PlanStateEdit
PlanStateEdit allows a bridged provider author to correct the diff generated by the TF
provider with additional information: the original Pulumi encoded inputs. This allows
Pulumi to recover from information lost in the Terraform encoding.

One example of this information loss is pulumi/pulumi-gcp#2372, where SDKv2's
representation of an unknown value `""` is not distinguished from an actual empty
string. A sister PR uses this capability to resolve pulumi/pulumi-gcp#2372.