Detailed diff in cross tests (#2366)
This exposes the detailed diff in the cross-test interface for both TF
and Pulumi allowing us to write tests which assert on that.

Also adds a cross-test for
https://github.com/pulumi/pulumi-terraform-bridge/pull/1696 for the
existing `TestChangingMaxItems1FilterProperty`. This is one of the cases
where the detailed diff is load-bearing for the diff decision. The cross
test verifies that TF also finds a diff here, so our core diff algorithm
must be making the wrong decision.

Related to https://github.com/pulumi/pulumi-terraform-bridge/issues/2293