Add SDKv2 PlanStateEdit (#2417)
PlanStateEdit allows a bridged provider author to correct the diff
generated by the TF
provider with additional information: the original Pulumi encoded
inputs. This allows
Pulumi to recover from information lost in the Terraform encoding.

The ability to recover information from Pulumi's representation is only
available outside the normal TF information flow. This is why diff
customizers can't fill the role of PlanStateEdit.

One example of this information loss is pulumi/pulumi-gcp#2372, where
SDKv2's
representation of an unknown value `""` is not distinguished from an
actual empty
string. A sister PR uses this capability to resolve (not this one GH)
pulumi/pulumi-gcp#2372.