Add more shim representation tests (#2456)
This PR simply adds test cases around tfgen and shim representations of
various schemata found in SDKv2 and PF. For PF I went through the docs
to cover each major advertised case. What's of special interest is
pinning down when Pulumi generates object types.

It seems to be happening in two cases:

1. when TF specifies object types directly in PF: 
       - Single Nested Attribute
       - Object Attribute
       - Single Nested Block

2. when TF specifies a block with MaxItems=1 that is a list or set in TF
but is flattened to an object in Pulumi:
- In SDKv2 this is typically a list-nested block with MaxItems=1, very
common
- In PF this may be a block with List Validators indicating that at most
1 item is supported; seems rare


It seems helpful to pin down what shim.Schema captures in each of these
cases.