Expose `crosstests.Create` (#2501)
This begins to turn cross-tests into a library with a defined interface. Existing
cross-tests say that they are a prototype level API, and they feel that way. The
prototyping was successful, and cross-tests have proven their worth. We should begin to
move towards a more discover-able API. The first function of which is:

```go
// crosstests/create.go

func Create(
	t T, resource map[string]*schema.Schema, tfConfig cty.Value, puConfig resource.PropertyMap,
	options ...CreateOption,
)
```

We can continue to iterate on this API, but we should begin to formalize a boundary
between the implementation of cross-tests and tests written with cross-tests.