[dynamic] Support delete on sdkv1 based providers (#2513)
Explicitly set `PlannedState: null` instead of leaving it empty.

As described by
https://github.com/hashicorp/terraform-plugin-framework/blob/ce2519cf40d45d28eebd81776019e68d1bddca6f/internal/fwserver/server_applyresourcechange.go#L63,
PF allows Null (not just nil) values to be used to send a delete. SDKv1
providers *always* attempt to deserialize planned state. This is fine if
it's null, but hits a nil pointer panic when it is nil.

Fixes https://github.com/pulumi/pulumi-terraform-provider/issues/36