Refactor detailed diff v2 to short circuit on nulls and unknowns (#2496)
This PR changes the logic in the detailed diff v2 calculation to
short-circuit on encountering nils and unknowns.

Previously, when we encountered a null or unknown in either news or olds
we would still recurse down and compare the non-nil values versus a nil
value in order to get any replacements which might be happening. We
would then simplify the diff later to trim these extra diffs but would
propagate the replacement to the right level.

We would now instead short-circuit on encountering a null or unknown and
walk the non-null value to find any properties which can trigger a
replacement. This makes it much easier to handle sets in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2451 as recursing
into sets is not necessary, as they are only compared by hashes.


This change is not meant to change any behaviour and is tested in
https://github.com/pulumi/pulumi-terraform-bridge/pull/2405

Stacked on https://github.com/pulumi/pulumi-terraform-bridge/pull/2515
and https://github.com/pulumi/pulumi-terraform-bridge/pull/2516