Improve set detailed diffs (#2451)
This change adds improved TF set handling to the detailed diff v2. The
main challenge here is that pulumi does not have native sets, so set
types are represented as lists.

### Diffing sets using the hash ###

To correctly find the diff of two sets we calculate the hash of each
element ourselves and do the diffs based on that. What makes this
somewhat non-trivial is that due to MaxItemsOne flattening we can't just
hash the pulumi `PropertyValue`s given to us by the engine. Instead we
use `makeSingleTerraformInput` to transform the values using the schema.
We then use the hashes of the elements in the set to calculate the
diffs. This allows us to correctly account for shuffling and duplicates,
matching the terraform behaviour of sets.

When returning the element indices back to the engine, we need to return
them in terms of oldState and newInputs because the engine does not have
access to the plannedState (see
https://github.com/pulumi/pulumi-terraform-bridge/issues/2281). To do
that we keep the newInputs and match plannedState elements to their
newInputs index by the set hash. Note that this is not possible if the
set element contains any computed properties - these can change during
the planning process, so we do not attempt to match and print a warning
about possibly inaccurate diff results instead.


### Unknowns in sets ###
Note that the terraform planning process does not allow a set to contain
any unknowns, because that breaks the hashing. Because of that plan
should always return an unknown for a set which contains any unknowns.
This accounts for cases where resolving the unknown can result in
duplicate elements.

Unknown elements in sets - the whole set becomes unknown in the plan, so
the algorithm no longer works. Currently we return an update for the
whole set to the engine and it does the diff on its side.

### Testing ###
This PR also includes tests for the set behaviour, both unit tests for
the detailed diff algorithm and integration tests using pulumi programs
for:
- Single element additions, updates and removals
- Shuffling, also with additions, updates and removals
- Multi-element additions, updates and removals
- Unknowns

### Issues ###

Builds on https://github.com/pulumi/pulumi-terraform-bridge/pull/2405

Stacked on https://github.com/pulumi/pulumi-terraform-bridge/pull/2515,
https://github.com/pulumi/pulumi-terraform-bridge/pull/2516,
https://github.com/pulumi/pulumi-terraform-bridge/pull/2496 and
https://github.com/pulumi/pulumi-terraform-bridge/pull/2497

fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2200
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2300
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/1904
fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/186