Enable TestConfigureSimpleValues
I'm unable to get the flake on my computer, even with `--count 1000`. The test wasn't
previously running in parallel, so it shouldn't be an interaction with other tests.

I've added better asserts so we can see which result is unexpectedly nil. I think our best
chance at fixing the test is to turn the test back on and then re-engage if/when it flakes
again.

Fixes https://github.com/pulumi/pulumi-terraform-bridge/issues/2530