Translate 'Terraform Workspace' to 'Pulumi Stack' (#2554)
This came up while running autogen for index files in Confluentcloud.

The top-level example had the following text:

```terraform
# Configure the Confluent Provider
terraform {
  required_providers {
    confluent = {
      source  = "confluentinc/confluent"
      version = "2.8.0"
    }
  }
}

# Option #1: Manage multiple clusters in the same Terraform workspace
provider "confluent" {
  cloud_api_key    = var.confluent_cloud_api_key    # optionally use CONFLUENT_CLOUD_API_KEY env var
  cloud_api_secret = var.confluent_cloud_api_secret # optionally use CONFLUENT_CLOUD_API_SECRET env var
}
```

It turns out that multiple provider examples in Confluentcloud have this
text in them, and descriptions and examples sections are elided. With
this change, we will gain more examples in at least that provider, but
potentially other providers as well. See pull request:
https://github.com/pulumi/pulumi-confluentcloud/pull/616